# openapi.yaml
openapi: 3.0.0
info:
  title: SportsDeck API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/reports:
    post:
      summary: Report a thread, post, or reply as inappropriate
      tags:
        - Reports
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contentType
                - contentId
                - reason
              properties:
                contentType:
                  type: string
                  enum:
                    - THREAD
                    - POST
                    - REPLY
                  example: POST
                contentId:
                  type: string
                  example: clxpost001
                reason:
                  type: string
                  example: This post contains hate speech.
      responses:
        '201':
          description: Report submitted
        '400':
          description: Missing fields, invalid contentType, or reporting own content
        '401':
          description: Unauthorized
        '404':
          description: Reported content not found
        '409':
          description: Duplicate report from same user
        '500':
          description: Internal server error
...
  /api/threads/{id}/sentiment:
    get:
      summary: Get AI-powered sentiment analysis of thread comments
      tags:
        - Threads
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: clxthread001
      responses:
        '200':
          description: Sentiment breakdown including overall and per-team sentiment
        '404':
          description: Thread not found or hidden
        '500':
          description: Internal server error
...
  /api/translate:
    post:
      summary: Translate a post, reply, or raw text to English
      tags:
        - Translate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  example: Hola, ¿cómo estás?
                contentType:
                  type: string
                  enum:
                    - POST
                    - REPLY
                  example: POST
                contentId:
                  type: string
                  example: clxpost001
      responses:
        '200':
          description: Translation result with original and translated text
        '400':
          description: Invalid contentType or no text provided
        '401':
          description: Unauthorized
        '404':
          description: Post or reply not found
        '500':
          description: Internal server error
...